(maint) - Compile with GCC
Had to remove -Wextra to get us compiling with GCC. JSONCPP
raises a “enumeral and non-enumeral type in conditional expression”
warning that cannot be dismissed with -Wno-enum-compare.

This is a cop out, but I don't have time to dig into the JSONCPP internals
right now.

The change in main.cpp was made because of the non-const
instantiation off of an rvalue. Sometimes = is still the right thing to do.