(CTH-108) Add DataContainer abstraction
This commit removes the direct use of JSON through out the code base and
replaces it with a DataContainer class.

This class is an abstraction over (currently) Jsoncpp and will allow us
to replace the underlying json implementation without having to touch
the entire code base.

The DataContainer class is implemented using variadic templates and
allows us to store basic scalars, typed vectors and recursive instances
of itself. See src/data_continer.h for documentation.

For now the DataContainer class is extended by the Message class. When
the reworked message format is implemented the Message class will use
DataContainer objects to implement header and data information.

This commit also introduces the use of Catch as a unit testing
framework.