(CTH-108) Add message abstracting
In the past we have been passing raw JsonCPP objects around in the
codebase, making it much harder for us to switch from one json library
to another as well as making it hard for us to change the message
format, since we're forced to update functions all over the place.

This commit introduces a new Message class which abstracts away the json
implementation. It allows us to change json libaries and update the
message format in the near future in a single location.

The commit precedes the codebase being updated to use the Message class.
We will pull the switch in the very near future.

This commit also introduces the first set of unit tests using the
Catch framework. We will continue to evaluate it and replace gtest with
it if it holds up.
https://github.com/philsquared/Catch