(CTH-106) Replace jsoncpp with rapidjson
This commit replaces the use of jsoncpp is the DataContainer abstraction
with rapidjson, standardising with out other C++ projects.

The vendored implementation of jsoncpp is removed and instead we vendor
rapidjson.

All jsoncpp includes have been removed.

valijson has been updated to current master and patched with
https://github.com/richardc/valijson/commit/1d3294f3169f278c937c2a6cd395646d126bda6d
to allow interaction with rapidjson.

This commit temporarily disables the inventory module due to cfacter's
incorrect forward declaration of rapidjson::Document.