(CTH-76) resolve external modules dir path
Use argv[0] to obtain the root of the cthun-agent dir tree and establish
the path of the external modules dir.

Note that this fix will not work in case argv[0] does not give the
entire path of the executable (this is up to the operating system) or in
case cthun-agent is invoked by a symlink.

Also, with this fix, the program will not break if it fails to determine
the path of the modules dir; in that case it won't try to load any
external module.