(CTH-128) Refactor the action invocation mechanism
ExternalModule is now in charge of creating the thread for the execution
of delayed tasks. Single path of execution for action invocation.

Delayed tasks now run in background; this will prevent cthun-agent to
immediately exit in case a task is still running. We'll fix that in
separate commit.

A possible invalid action output is now reported in the result files; we
should add a specific field for the action 'success' in the related data
schema.

An external_module_test.cpp test case now refers to the new async
ExternalModule::executeDelayedAction(), which spawns the delayed action
thread; this test requires a usleep() pause.