(CTH-144) Non-init Configuration returns default values
A non initialized Configuration object will not throw an exception in
case a known entry is requested via get().

If the Horsewhisperer::SetFlag fails, we now throw an exception.

Using a map for the default Entries, instead of a vector.
New error class for bad configuration entries.
Changing name of the 'configured_' flag to 'initialized_', to have a
more direct reference to the initialize() method.