(maint) Improve overhead tolerance in thread_container_test
Incrementing the value of a pause that allows a thread that should
finish immediately to actually finish.

Without this change, in case the thread_container_test was executed
by a machine with low CPU resources, the thread processing overhead
for the OS could have triggered a terminate call in different tests.