CTH-242 allow cthun-client to be already installed
Our use of cthun-client via an ExternalProject, though convenient,
causes issues when building in a more restricted environment as
provided when doing a cthun-agent-vanagon build via the vmpooler.
The -DCMAKE_TOOLCHAIN_FILE and additional configuration parameters
are not passed to the sub-cmake that the ExternalProject invokes,
so the build of cthun-client can fail to find build dependencies
like the C++ compiler.

Here we allow for the cthun-client library and headers to be already
installed and findable via `find_package(cthun-client REQUIRED)`
when -DEXTERNAL_CTHUN_CLIENT=ON is specified (the default), and
update the existing vendor/cthun-client.cmake to advance the same
`cthun-client_LIBRARY` and `cthun-client_INCLUDE_DIR` symbols that
the new cmake/Findcthun-client.cmake does.

Also we change the default build to be a Release build, but keep
the developer Makefile using a Debug build with the submodule version
of cthun-client.