(maint) Pass spool-dir as argument to Agent and RequestProcessor
Passing the spool-dir path as an argument of the Agent and
RequestProcessor ctors in order to: 1) allow dependency injection for
unit tests and avoid reconfiguring the Configuration singleton; 2)
reducing coupling with Configuration / HW; 3) being consistent to what
we do with certs and modules-dir.

Using Configuration wrappers to retrieve options instead of using HW
directly (in main.cc).