(CTH-240) Make modules configurable
Here we add the cli flag --modules-config-dir which specifies a location
for a collection of .cfg files, named after a module, which can be used
to configure.

Each file in the config dir will be read as a JsonContainer and stored
in the config singleton. The config dir will default to
/etc/puppetlabs/cthun-agent/modules.d as recommended by UX.

Here we also configure leatherman::logging to start up pointing at cout
and at info level. This allows us to warn about badly configured
module-conf-dirs at startup which is useful for now. We will have to
change this as we become closer to running as a service. I suspect there
are going to be a lot of changes to the Configuration class in the
future so we need to keep a close eye on this.