(CTH-241) Implement pxp-module-puppet
Here we implement the puppet pxp module. This module has two actions,
"metadata" and "run". The "run" action will attempt to start a Puppet
run with `puppet agent -t` on the local machine and like other modules
report the result on STDOUT.

Result is created based on the content of last_run_report.yaml, the
exitcode of `puppet agent -t` and may include a field "error" which is a
string explaining any error that may have occurred.

A README has been added explaining installation, configuration and usage
as well as the output from any error that could occur.