(CTH-345) Update default config file path
Here we change the default config file path to be
/etc/puppetlabs/pxp-agent/pxp-agent.cfg to better fit with puppet-agent.
We also stop looking for a config file in $HOMEDIR/.pxp-agent and update
the README to reflect reality.

With this commit we also enforce the need for a config file with the
goal of moving towards acting more like a real daemon.