(maint) Fix lookup for cpp-pcp-client library
On Windows, library lookup behaves a little differently. We put dynamic
libraries in a `bin` directory, and generally libraries don't use the
`lib` prefix and use `.dll`/`.lib` suffixes. For libraries that might be
loaded by Ruby we use Linux-style names because Ruby expects it. Fix
library lookup to allow for all of that.