Update PXP readme example config
Prior to this commit, the pxp readme said to connect to the pcp-broker
with the ending url pxp, however the example config files in pcp-broker
define the url entrypoint with pcp. To avoid confusion, this commit
updates that by changing the example config to have the server url end
in pcp.

It also updates the example cert files to match what you would expect to
find after installing puppet-agent on a system.