Update example with config instead of configuration
Prior to this commit, the example said to use 'configuration' as a key
when passing in a json hash to pupet run. The module puppet ruby binary
is actually looking for 'config', not 'configuration', so the run
command would fail due to the missing key. This commit changes that by
updating the readme example to use 'config' instead of 'configuration'.