(maint) Disable detailed exitcodes
This replaces -t (--test) with its implied command line flags except for
--detailed-exitcodes (also without --ignorecache which does literally
nothing). This is because, with detailed exitcodes enabled, puppet
doesn't distinguish between a failure to execute a puppet run and
resource failures during the run. Also, puppet exits 2 with detailed
exitcodes if there are resource changes, which was previously being
dealt with as a failure case.

With detailed exitcodes disabled, we can rely on the fact that an exit
code of 1 means that the puppet run failed to execute, while an exit
code of 0 means we should expect a report to be available that will
allow us to determine whether everything that happened *during* the run
was successful.