(PCP-30) Agent should be able to start unconfigured
Here we allow the agent to start in a unconfigured mode to help
facilitate acceptance testing.

When the key parameters server, ca, cert and key are not set,
instead of starting normally the server will start a busy loop. At this
point the agent should still behave normally to signals.