(PCP-28) Add interpreter field to module config options
Here we allow modules to specify an "interpreter" field in their config
files. If this field is set, the agent will use it to execute a module
action instead of assuming that the script is executable. This allows us
to get around having to use bat files and shebang lines on different
platforms.

We do this by slightly changing the api of ExternalModule. A second
ctor is added with the config as a parameter so we can check for the
interpreter field before validating the rest of the config.
validateAndSetConfiguration becomes validateConfiguration, since
configuration is set before the method is called.