(PCP-63) pxp-agent will always demonize when unconfigured
Here we change the when we start in unconfigured mode from "no config
file is specified, the default config file doesn't exist, drop all flags
and start daemonized" to "If options are not set that we require to
connect, start unconfigured". We will continue to fail in the case of
config errors that we cannot recover from, such as parsing invalid
flags, not being able to process a given config file due to bad values,
not being able to log and (in the future) not being able to write a
pidfile.