(PCP-6) Reopen log file after SIGUSR2 on *nix
Registering a handler for SIGUSR2 that will reopen the file associated
with the log file stream (new class var of Configuration), in case we
log on file on *nix platforms.

Adding new platform specific section on configuration.hpp and related
implementation files in src/configuration/posix  and /windows.
Adding reopenLogFile() method to Configuration.