(PCP-96) Remove unneeded config options and set correct defaults
Here we remove the need for the `interpreter` module config option and
the `puppet_bin` pxp-module-puppet config options.

`interpreter` has been removed completely. As a result, all modules must
be executable and not have a file extension on *nix systems. On Windows
all modules must have a .bat file with the same name as the module which
is responsible for calling the module correctly.

`puppet_bin` is still a valid configuration option, but will default to
the correct executable from the puppet-agent on both *nix and Windows.

This commit means that pxp-module-puppet can be called without needing a
configuration file.