(PCP-78) Add --pidfile option for *nix platforms
Here we allow specifying the path of the PID file on *nix.
Note that on Windows we don't write the PID to file; we ensure that only
one instance of pxp-agent is running at time by using a mutex.

Modifying the PIDFile interface.
Checking the pidfile path.
Updating unit tests.