(PCP-113) Failed puppet run should return more information
In the past if a Puppet run exited with a non-zero exit code we would
not attempt to inspect the last_run_report.yaml and return the default
set of unknown results.

Here we change this behaviour and inspect the last_run_report.yaml even
if the run exited with a non-zero code. If the "time:" entry is newer
than when than when the run was kicked we will populate the results with
the values from the last_run_report.yaml. If it is older, we return the
default unknown set of results.