(PCP-122) Don't set PATH in pxp-agent service scripts
Prior to this commit, some of the pxp-agent service
scripts were setting PATH. This appears to be
unneccesary. Additionally, the specific PATH that
was being set could lead to unexpected use of the
puppet-agent's "private" utilities, such as gem.

This commit removes PATH where it was set in the
pxp-agent service scripts.