(PCP-123) Add default flags to puppet agent run
Here we add the flags `--onetime`, `--no-daemonize` and `--verbose` to
all puppet agent runs by default. This is done by calculating the union
between the set of passed flags and the set of default flags so that if
a default flag was set by the caller, we do not set it again.