(PCP-119) Add machine readable error_type to response
Here we add a new field to the response, error_type. This field is
intended to be machine readable so that consumers of the response
message can easier determine how to react to failures.

We are also able to refactor #get_result_from_report by adding
a #make_error_result function. This simplifies it's calling convention
by removing the need for the optional error string and fail_hard
boolean. Calls that do not require the last_run_report to be parsed can
now directly call #make_error_result.