P-117) Fix racy checks in running? and disabled?
In the past it was possible for a run to start between checking if there
was a run and running.

Here we no longer check if the lock files exist and instead kick a run
regardless and inspect the output to determine if a run was already
underway or if puppet was disabled. We still report the same message
back to the consumer.