(PCP-174) Acceptance - expand install utils
This commit expands the acceptance install_utils in order to
support installation on the full range of platforms for
acceptance testing.

It also incorporates the Rakefile structure from puppet/facter/hiera
in order to allow Jenkins pipeline support for AIO. The Rakefile
retains usage and checks for SUITE_COMMIT and TEST_TARGET to ensure
that it continues to work with the existing Jenkins job parameters.