(PCP-136) Try to fix Catch failure with different macro
The PIDFile ctor test, section "it fails if the dir does not exist" was
causing an error on different platforms; apparently Catch was not
filtering the asserted error when the REQUIRE_THROWS macro was used.
Here we replace such macro with REQUIRE_THROWS_AS to investigate the
behaviour of the unit tests.