(PCP-129) Requests for internal modules must be blocking
With this commit we require that PXP requests for internal modules must
be of blocking type.

This limitation will semplify the processing workflow of non-blocking
requests; that's important since, for PCP-129, we will delegate the
child process to write directly on file the output of stdout/stderr
streams - to cope with that when processing the outcome of internal
modules we would have to implement a different path of execution.