(PCP-148) Change acceptance tests to use static host config files
To bring pxp-agent in-line with other Puppet agent components, change the Rakefile to use static config files.
Which host file to use is determined by TEST_TARGET in the environment. Rakefile exits if this is not set.
Also begin renaming the test setup term "packages" to "aio" because that's what it actually is.