(maint) Acceptance tests - add an explicit assertion in the event of SSL test failing
The acceptance test for invalid SSL config depends on a grep statement returning 0.
If this fails, we only see that grep failed; without seeing the text it was grepping.
This commit catches grep failing then fails on an exception that will compare the expected log entry with the actual log contents.