(maint) Remove ssh-agent call
Previously, acceptance was only failing in CI, but not when run locally,
and the failure was that the ca_crt.pem was never rsync'ed to the agent
node:

    WebSocket configuration error (ssl-ca-cert file '/root/test-resources/ssl/ca/ca_crt.pem' not found

I believe the root cause is due to using ssh-agent on the coordinator
node. This step was only necessary so we could git clone the private
pcp-broker repo. Now that the repo is public, the step is not necessary,
and it eliminates having to interact with ssh-agent on the shared CI
coordinator.

This commit also increases the beaker log level so we can see what
commands beaker is executing.