(maint) acceptance - replace rsync with scp_to
Previously, acceptance tests were failing in CI due to the lack
of a defined SSH key needed to execute rsync on the test runner.

This commit replaces the call to rsync with the beaker helper
method `scp_to`. Since, beaker manages the SSH keys as part of
the startup process, further mention of the SSH key is not
required.