(PCP-168) Synchronize status metadata writing
With this commit we synchronize reading and writing operations on
transaction status metadata files by non-blocking tasks and status
requests handlers.

The status query action now reads the PID file before accessing the
metadata file. In case the PID is related to an existent process, the
status handler will sleep for 0.1 s, to ensure that the non-blocking
task have the time to acquire the lock that enables writing the metadata.

Both the non-blocking task and status handler use a boost named_mutex to
synchronize the access to the metadata file. Such mutex will be named
after the first 8 letters of the transaction id of the interested
non-blocking task.