(PCP-172) Synchronize access to status metadata file
With this commit we synchronize reading and writing operations on
transaction status metadata files by non-blocking tasks and status
requests handlers.

The status query action now reads the PID file before accessing the
metadata file. In case the PID is related to a non-existent process, and
its mutex is still in the cache, the status handler will sleep for 0.1 s,
to ensure that the non-blocking task has the time to lock the mutex
and update the metadata file.

The synchronization relies on the ResultsMutex singleton that caches
mutexes named against transaction IDs.