Make agent run more robust
Check if the agent disabled lockfile already exists before running the agent

Before running create a disabled lockfile containing a message that pxp run is
in progress. Ensure this file is unlinked afterwards.

Use a pxp specific agent disabled lockfile to prevent further runs by the
daemon, cronjobs or manually. The others will get a nice message that a pxp run
is in progress.

If a puppet run is already in progress, try to wait some time. max waittime is
60s now. this should be configurable.