(PCP-176) Add support for 32bit Puppet on 64bit Windows
Puppet and pxp-agent will be in the Program Files x86 folder if Puppet is 32bit on 64bit Windows.
To do this, I've added a helper method to get the path to pxp-agent.
Removed the static strings in the pxp_agent_version test and replaced with a call to the helper.