(PCP-174) Acceptance - add OSX root paths to helper
This commit adds the `/var/root` base paths to the
pxp-agent acceptance helper. It also refactors the
helpers to accommodate OSX as an option when building
paths for cert installation.

Prior to this commit, acceptance tests on OSX failed due to the
mismatch in the expected (`/root`) and actual (`/var/root`) of
the base path for the root user.

[skip ci]