(maint) Acceptance - remove incorrect named arguments from SSL config test
The invalid_ssl_config test was calling a helper method with what appeared to be a named argument, but was actually just creating a local variable.
Note that proper keyword arguments are possible but are only in Ruby 2.0 onwards; and our CI might use Ruby 1.9.3 so they have not been used here.