(maint) Acceptance tests should give test failures and not exceptions when checking log file
A couple of our tests use beaker retry_on to wait for a log file to contain an expected log entry.
retry_on failing results in the test being marked as 'error' and no indication of what the file actually contained.
This commit adds a general helper method that waits for a file to contain an expected string using retry_on; then in the case of it failing, the method will rescue then call an assertion.
The test will then be marked 'failed' and the test output will show the actual content of the file.