(PCP-237) Normalize modules-dir and modules-config-dir values
With this commit we expand the paths of the modules and modules
configuration directories, when validating Configuration options.

Creating modules config dir for Configuration unit tests (in
lib/tests/resources), as pxp-agent assumes the existence of such dir (as
per Puppet specs) and, with this commit, it verifies if the path
corresponds to a existent dir.