(PCP-251) support for saving stdout/stderr/exitcode to files
Previously the the pxp-module-puppet (or in general every module)
was supposed to pass it's output on the stdout which was a pipe
to the pxp-agent process. This din't work well in case the pxp-agent
proces died / was restarted.
So newly when the module is passed the "output_files" entry in the
argumets, it is required to save its stdout/stderr/exit code into
the respective files listed in the "output_files" entry. This
way the module process is much better decoupled from the pxp-agent
process.