(PCP-207) Ext modules interface: rely only on files
We're simplifying the interface for external modules by requiring that
the output must be written only to file in case the `output_files`
entry is included in the action input.

This change implies that pxp-agent will need to read the outcom of the
action run from file.

On the other hand, modules will have a simpler output requirement, as
they won't need to guarantee the same output on files and stream.

[skip ci]