(PCP-208) ThreadContainer stores the name as well
When adding a thread objecty to a ThreadContainer instance, the user must
also specify the name of the thread. This enables inspection.

Using an unsorted_map instead of a vector to store the thread instances;
this enables the retrieval by name.

Raising an Error in case ThreadContainer::add is called for an existent
name.

Adding find() and getThreadNames() methods.

Making const member functions where possible.

Adding new unit tests.

Updating RequestProcessor to comply with the new interface and with the
possibility of having stored a thread with the same name (using a mutex
for that). For a given non-blocking request, it's execution thread will
be named against the transaction_id.