(maint) Acceptance tests need to allow time for agent association with broker
PCP-167 changed the tests to check assocation with pcp-broker instead of grepping pxp-agent log for association log entries.
This occasionally fails in CI as the agent does not appear as associated with the broker.
This commit adds retry logic so that the pcp-broker inventory is checked a number of times before failing.

[skip ci]