(PCP-208) New data structures for processing actions
With this commit we introduce new data structures to allow processing
the results of actions when they are retrieved from file, in the same
way as when they are obtained directly at the end of an action run. This
will enable pxp-agent to be fully resilient to restarts occurred during
an action run.

Also, the changes introduced here enable a better encapsulation of the
logic for formatting messages and provide new metadata information about
the processing of tasks. All that will ease the implementation of the
PXP changes suggested in PCP-202.

- removing the ActionOutput data structure; the results and outcome of
  an action run are now contained in the new ActionResponse that
  includes 1) ActionOutput (the action's output on stdout, stderr, and
  exit code) and 2) an action_metadata JSON object (which schema is
  enforced);
- refactoring the logic for the validation of the action output in order
  to make it accessible to the 'status query' implementation;
- creating a new ResultsStorage class that provides an interface to
  access and update the data stored in the spool directory;
- refactoring the module and request types.