(maint) Don't modify metadata file at every unit test run
We must update the metadata file in tests/resources used to test the
purge functionality. Previously, we were leaving the file in the new
state, needed by the unit test. As a consequence, the file change would
be inserted in a git commit everytime "git add -A" was invoked after a
pxp-agent-unittests run. We now prevent this by restoring the metadata
file state after the relevant ResultsStorage unit test.