(PCP-317) flush pxp-module-puppet output explicitly
There was a race condition between creating the exitcode file (which
pxp-agent uses to determine if a pxp-module-puppet execution finished)
and the actual exit of the pxp-module-puppet process (which implicitly
flushes its output). This could lead (and led) to pxp-agent reading
incomplete output of the pxp-module-puppet.
To prevent this we flush the output of pxp-module-puppet explicitly
before writing the exitcode file.