(PCP-306) run report in tempfile
Before this commit we let puppet runs executed from px-module-puppet write
their run reports to the default location. Since the file is overwritten
on every puppet run, we could never be sure when reading the file whether
it was produced by the puppet run triggered by pxp-module-puppet or by
a concurrent puppet run.
To avoid this, we update the pxp-module-puppet in this commit such that
it ensures (by passing the --lastrunreport option) that the puppet runs
it triggers write their run reports to uniquely named temporary files
rather than to the default location.