(PCP-228) acceptance: run Puppet and expect Failure
This change adds a test using pxp-agent-puppet module to run puppet when
we expect failure and ensure it is captured properly.
We also add a library to acceptance to create a tmp environment, linked
into environment_path and a teardown is registered.
Note that the method for environmentpath is memoized.  So if you add
something to the default environmentpath on master, you should not use
this method.

[skip ci]