(maint) Fix up AppVeyor
AppVeyor images were updated to include software that installs OpenSSL
binaries to `C:\Windows\System32`.
https://msdn.microsoft.com/en-us/library/7d83bc18.aspx describes the
Windows search path, which puts the Windows system directory before
PATH. The easiest way to ensure the libraries are loaded earlier is to
put them in the same directory as the test executable, so do that.