(PCP-349) support running puppet as non-root
This commit makes it possible to run puppet under a non root user.
Previously this was not possible because puppet failed to expand
the user home directory. This was due to the environment cleaning
performed by the `Puppet::Util::Execution.execute` function used
for the puppet run execution which included unsetting of the HOME
environment variable.
This patch solves that by fixing the environment (re-adding the
necessary environment variables to it) after the cleaning. The
values of the re-added variables are obtained by reading the pwentry
corresponding to the current effective user id.