(PCP-423) Add `allowed-keepalive-timeouts` option
Adds the `allowed-keepalive-timeouts` option, with a default of 2. The
option is used to decide how many consecutive keepalive responses can be
lost before closing the connection and trying to reconnect. With the
default of 2, the connection will be closed after 3 lost keepalive
responses; the agent will then attempt to reconnect to the broker and/or
failover broker.