(PCP-452) Add option to configure Association timeout
The `association-timeout` option is now used to set the timeout for the
overall Association process (default to 15 s).

The old `association-timeout` (ttl for the Association request) can now
be specified by the `association-request-ttl` option (default to 10 s).

Renaming the `pcp-message-timeout` option to 'pcp-messag-ttl', for
consistency with the other ttl option.