(PCP-472) Ignore pxp-module-puppet `env` input arg
Ignoring the `env` array to avoid letting the user to set custom
environment variables as we currently don't have a use case for that.

The `env` entry of the JSON metadata's input object remains defined
for backwards compatibility, but it's not a required entry anymore.
Thus, pxp-module-puppet interface remains the same; in the future we may
want to completely remove the `env` entry, but that would require
bumping the interface version.

Updating unit tests.

Updating docs.