(maint) Fix path for pxp-agent using new Windows MSI layout
The Windows MSI layout changed to sharing binaries for C++ applications.
That means pxp-agent.exe now relies on files in puppet/bin. Update the
test to include that in PATH.

For now hard-code in the test. It's not trivial to amend the
privatebindir in config/aio/options.rb, since the setting is
platform-dependent. This might eventually live in Beaker, but since it's
only needed for this test that may not be worth it.

[skip ci]