(PCP-545) Force Puppet output to UTF-8 when unknown
When Ruby's default encoding is US-ASCII, Puppet may still use other
character encodings from the manifest. Request that characters are
printed as UTF-8 (to hopefully handle any encodings Puppet may deal
with). Later, try to change all execution output to UTF-8 (because
that's what we want to compare against with regexes anyway), and if that
fails assume we have UTF-8 characters but used a non-UTF-8 locale.

There may be some un-handled corner-cases, but it shouldn't make them
worse. If you try to run with a German codepage on Windows, but a
manifest has UTF-8 characters that can't be represented in it, you
already have a problem.