(PCP-549) Wait for other Puppet runs to finish
When `pxp-module-puppet run` is triggered and Puppet was already
running, the module would previously identify the failure and return an
error message saying Puppet could not be run because another run was in
progress. This leads to a potentially slow feedback cycle of retrying
Puppet until it can run, when all the module user wants is to ensure a
Puppet run happens with the arguments they specified.

Change pxp-module-puppet to wait until the prior Puppet run completes
and attempt to run again until it succeeds or encounters a different
error.