(PCP-557) Ensure broker runs with internal mirrors
Periodically starting the broker will take a long time. It appears to be
downloading dependencies when `lein tk` is run, despite `lein deps`
being previously run. This is probably because we ran `lein with-profile
internal-mirrors deps`, but then run `lein tk` with a different profile.
Update to run `lein with-profile internal-mirrors tk`.

[skip ci]