(PCP-572) Update how acceptance terminates sleep processes
The method of ending a running sleep process using SIGALRM does not work on Solaris 10.
We don't need SIGALRM specifically, we just need 'sleep' to terminate; so this commit sends SIGTERM instead of SIGALRM.

Because the Solaris failure caused knock-on failures in other tests; this commit also adds a teardown to the tests that make use of sleep/kill.

[skip ci]