(PCP-601) Check for flags with leading whitespace
Previously it was possible to fool the security checks in pxp-module-puppet
(which are supposed to verify whether the passed puppet flags are acceptable
or not) by prefixing the unacceptable flags with whitespace. When such a flag
was passed, it was accepted and was even effective due to the way the puppet
command to execute is composed.
This patch makes this exploit impossible by stripping the whitespace from the
flags before performing the checks.