(PCP-607) Delete acceptance test run_puppet_twice.rb
The functionality to have a Puppet run action wait on a current agent run to complete is not yet on the stable branch.
Therefore having the acceptance test for it on stable is wrong - this commit deletes it.
PCP-607 still requires work - this mistake was not noticed because the test does not fail in most cases.

(skip ci)