(PCP-601) Limit arguments to pxp-puppet flags to needed characters
Previously anything that didn't start with "--" was treated as a flag
argument rather then a flag and not validated. This allowed the flag
whitelist to be bypassed by submitting flags that started with
characters other than "--" but were coerced down to "--".  After this
commit all arguments are limited to only the characters necessary to
express the whitelisted flags and valid arguments to those flags.