(PA-654) Acceptance - ensure puppet is not running as a service
PA-654 is a bug in puppet-agent where puppet is enabled as a service after package install.
This caused pxp-agent acceptance to fail due to its test that restarts the agent, followed by
its test that depends on checking the running puppet agent instances.

This commit adds a step to pre-suite that ensures puppet service is stopped and disabled.

[skip ci]