(PCP-631) Prevent hanging on Windows ssh after shutdown
The beaker restart helper sends shutdown to Windows, but then hangs for
a long timeout as ssh doesn't close cleanly. This is filed as BKR-971.

As a work-around, add a 5s timeout after sending shutdown to allow it to
begin, then close the SSH connection and start attempting to reconnect.

[skip ci]