(PCP-632) Fix run_puppet_twice.rb on OSX
This test fails in CI when it requires either 1 or 2 puppet agent processes to be running but discovers 3.
1 instance is from puppet running as a service (despite the pre-suite stopping this).
Another instance is from puppet agent being used for another purpose (puppet agent --configprint).

This commit updates the test to not fail immediately on finding 3 pids and also adds a workaround in shutting down puppet at the start of the test.

[skip ci]